requirements.txt
dm-haiku
optax
einops
chex==0.1.5

numpy
datasets
tokenizers

wandb
pyyaml

mypy
flake8
isort
black
pytest
